Upgrade to clap 4

+++ src/args.rs
@@ -1,20 +1,20 @@
-use clap::{Parser, Subcommand};
+use clap::{ArgAction, Parser, Subcommand};
 use std::net::SocketAddr;
 use std::path::PathBuf;
 
 #[derive(Debug, Parser)]
 pub struct Args {
     /// Turn debugging information on
-    #[clap(short, long, global = true, parse(from_occurrences))]
-    pub verbose: usize,
-    #[clap(subcommand)]
+    #[arg(short, long, global = true, action(ArgAction::Count))]
+    pub verbose: u8,
+    #[command(subcommand)]
     pub subcommand: SubCommand,
 }
 
 #[derive(Debug, Subcommand)]
 pub enum SubCommand {
     Bait(Bait),
-    #[clap(subcommand)]
+    #[command(subcommand)]
     Infect(Infect),
 }
 
@@ -22,14 +22,14 @@ pub enum SubCommand {
 #[derive(Debug, Clone, Parser)]
 pub struct Bait {
     /// Address to bind to
-    #[clap(short = 'B', long, env = "SH4D0WUP_BIND")]
+    #[arg(short = 'B', long, env = "SH4D0WUP_BIND")]
     pub bind: SocketAddr,
     /*
     /// The upstream server to reverse proxy to
-    #[clap(short = 'U', long)]
+    #[arg(short = 'U', long)]
     pub upstream: Option<String>,
     /// Do not modify Host: and similar headers
-    #[clap(short = 'K', long)]
+    #[arg(short = 'K', long)]
     pub keep_headers: bool,
     */
     /// Path to the plot to execute
@@ -53,10 +53,10 @@ pub struct InfectPacmanPkg {
     /// Where to write the modified package to
     pub out: PathBuf,
     /// Update a key in .PKGINFO (a key can be set multiple times)
-    #[clap(long)]
+    #[arg(long)]
     pub set: Vec<String>,
     /// The command to inject into the package that's executed once during install
-    #[clap(short = 'c', long)]
+    #[arg(short = 'c', long)]
     pub payload: String,
 }
 
@@ -68,10 +68,10 @@ pub struct InfectDebPkg {
     /// Where to write the modified package to
     pub out: PathBuf,
     /// Update a key in ./control
-    #[clap(long)]
+    #[arg(long)]
     pub set: Vec<String>,
     /// The command to inject into the package that's executed once during install
-    #[clap(short = 'c', long)]
+    #[arg(short = 'c', long)]
     pub payload: String,
 }
 
@@ -83,7 +83,7 @@ pub struct InfectOci {
     /// Where to write the modified package to
     pub out: PathBuf,
     /// The command to inject into the package that's executed once during install
-    #[clap(short = 'c', long)]
+    #[arg(short = 'c', long)]
     pub payload: String,
 }
 
@@ -95,9 +95,9 @@ pub struct InfectApkPkg {
     /// Where to write the modified package to
     pub out: PathBuf,
     /// Update a key in .PKGINFO (a key can be set multiple times)
-    #[clap(long)]
+    #[arg(long)]
     pub set: Vec<String>,
     /// The command to inject into the package that's executed once during install
-    #[clap(short = 'c', long)]
+    #[arg(short = 'c', long)]
     pub payload: String,
 }
